% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors-data.R
\docType{data}
\name{contributors}
\alias{contributors}
\title{Data of contributors for each issue of Blast, Egoist, Freewoman, New Freewoman, and Poetry Magazine.}
\format{a data.frame object}
\source{
\href{https://sourceforge.net/projects/mjplab/files/}{Modern Journals Project}
}
\usage{
data(contributors)
}
\description{
This metadata can be found at https://sourceforge.net/projects/mjplab/files/
}
\keyword{datasets}

