% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_vec-data.R
\docType{data}
\name{date_vec}
\alias{date_vec}
\title{Date published data for each issue of Blast, Egoist, Freewoman, New Freewoman, and Poetry Magazine.
Taken by using strip_dates to identify date published.
This metadata can be found at https://sourceforge.net/projects/mjplab/files/}
\format{a data.frame object}
\source{
\href{https://sourceforge.net/projects/mjplab/files/}{Modern Journals Project}
}
\usage{
data(date_vec)
}
\description{
Date published data for each issue of Blast, Egoist, Freewoman, New Freewoman, and Poetry Magazine.
Taken by using strip_dates to identify date published.
This metadata can be found at https://sourceforge.net/projects/mjplab/files/
}
\keyword{datasets}

