% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_model.R
\name{graph_model}
\alias{graph_model}
\title{Graph results of LDA}
\usage{
graph_model(df, baseColNames = c("text", "date", "id", "journal"),
  chunkLen = 10)
}
\arguments{
\item{df}{data.frame a trained LDA model, taken from train_model}

\item{baseColNames}{vector string names of the "base" columns - text, date, id, and journal}

\item{chunkLen}{the number of graphs per page. If less than the number of topics, which is found
automatically, splits the dataframe into n.topics/chunkLen sections, each of which has at most
chunkLen plots of topic makeup over the documents}
}
\value{
n.topics graphs spread of chunkLen pages
}
\description{
Graph results of LDA
}
\examples{
date_vec <- date_vec
dataframe <- stripped_xmlData
marsden_journals <- make_model(dataframe, 10, journalVec = c("Blast", "Egoist", "Freewoman", "NewFreewoman"))
graph_model(marsden_journals, chunkLen = 6)
}
\seealso{
\code{\link{train_model}}, which runs LDA on a Mallet instance
}

