% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLDAvis.R
\name{makeLDAvis}
\alias{makeLDAvis}
\title{Easy interface to LDAvis, only requires a trained topic.model as input}
\usage{
makeLDAvis(topic.model, outDir = tempfile(), openBrowser = TRUE,
  asGist = FALSE, ...)
}
\arguments{
\item{topic.model}{Mallet-object a trained topic.model from MALLET}

\item{outDir}{character directory to store html/js/json files}

\item{openBrowser}{logical should R open a browser to create visualizations?}

\item{asGist}{logical should the vis be uploaded as a gist? Requires gistr}

\item{...}{arguments passed onto gistr::gist_create}
}
\value{
html/js/json files of a LDAvis visualization
}
\description{
Easy interface to LDAvis, only requires a trained topic.model as input
}

