% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelToDataframe.R
\name{modelToDataframe}
\alias{modelToDataframe}
\title{Run LDA on a Mallet instance}
\usage{
modelToDataframe(topic.model, prepped_corpus, num.topwords = 5,
  textcolname = "text", incCorpusText = TRUE)
}
\arguments{
\item{topic.model}{topic.model usually taken from make_model, a Mallet object}

\item{prepped_corpus}{data.frame usually taken from prep_corpus, a data.frame with fields for text, journal, date}

\item{num.topwords}{numeric number of words to display on topic labels}

\item{textcolname}{string the name of the text column from prepped_corpus}
}
\value{
a data.frame of topic frequencies by document that we can analyze and graph
}
\description{
Run LDA on a Mallet instance
}
\examples{
date_vec <- date_vec
dataframe <- stripped_xmlData
marsden_journals <- make_model(dataframe, 10, journalVec = c("Blast", "Egoist", "Freewoman", "NewFreewoman"))
marz.df <- train_model(marsden_journals, dataframe, c("Blast", "Egoist", "Freewoman", "NewFreewoman"))
}
\seealso{
\code{\link{prep_corpus}}, which creates the data.frame this function draws on
\code{\link{make_model}}, which creates the mallet object this function uses
}

