% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_corpus.R
\name{prep_corpus}
\alias{prep_corpus}
\title{Create a database of text and metadata to feed into make_model}
\usage{
prep_corpus(xmlfolder, date_vec, wordsToRemove = NA, stemDoc = NA,
  pattern = ".tei.xml")
}
\arguments{
\item{xmlfolder}{string input folder with text files to run. Typically contains either XML files
from the MJP or files cleaned with stripXML}

\item{date_vec}{vector vector of dates, usually from strip_dates}

\item{wordsToRemove}{vector a vector of words to remove from the corpus; an initial stopword list}

\item{stemDoc}{logical if TRUE, runs SnowballC's stemDocument function}

\item{pattern}{string pattern of the endings of the files of xmlfolder; defaults to
'.tei.xml', but if xmlfolder is taken from stripXML, should be '.txt'}
}
\value{
a data.frame of text for running with MALLET
}
\description{
Create a database of text and metadata to feed into make_model
}
\examples{
dataframe <- stripped_xmlData
}
\seealso{
\code{\link{strip_dates}} and \code{\link{stripXML}} which this depends on, as well as 
\code{\link{make_model}}, which takes the data.frame created by this function as input
}

