% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_dates.R
\name{strip_dates}
\alias{strip_dates}
\title{Creates a vector of publication dates given a folder of files from the Modern Journals Project}
\usage{
strip_dates(xmlfolder, pattern = ".xml")
}
\arguments{
\item{xmlfolder}{input folder of files to parse dates from}

\item{pattern}{file endings of the files. Defaults to ".xml"}
}
\value{
POSIXCT vector of date published for each of the files
}
\description{
Creates a vector of publication dates given a folder of files from the Modern Journals Project
}

