% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripped_xmlData-data.R
\docType{data}
\name{stripped_xmlData}
\alias{stripped_xmlData}
\title{Loads the text of parsed XML files for each issue of Blast, Egoist, Freewoman, New Freewoman, and Poetry Magazine.
This parsing, or removal of HTML, was done using the function stripXML
This metadata can be found at https://sourceforge.net/projects/mjplab/files/}
\format{a data.frame object}
\source{
\href{https://sourceforge.net/projects/mjplab/files/}{Modern Journals Project}
}
\usage{
data(stripped_xmlData)
}
\description{
Loads the text of parsed XML files for each issue of Blast, Egoist, Freewoman, New Freewoman, and Poetry Magazine.
This parsing, or removal of HTML, was done using the function stripXML
This metadata can be found at https://sourceforge.net/projects/mjplab/files/
}
\keyword{datasets}

