% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlData-data.R
\docType{data}
\name{xmlData}
\alias{xmlData}
\title{Loads the text of XML files for each issue of Blast, Egoist, Freewoman, New Freewoman, and Poetry Magazine.}
\format{a data.frame object}
\source{
\href{https://sourceforge.net/projects/mjplab/files/}{Modern Journals Project}
}
\usage{
data(xmlData)
}
\description{
This metadata can be found at https://sourceforge.net/projects/mjplab/files/
}
\keyword{datasets}

