% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model.R
\name{make_model}
\alias{make_model}
\title{Create a Mallet object; from here we can gather word frequency statistics, and feed into createLDAvis}
\usage{
make_model(dataframe, n.topics, idcolname = "id", textcolname = "text",
  datecolname = "date",
  stopListFile = "~/Dropbox/The Egoist PDFs/R/blankstopwords.txt",
  optFreq = 20, burnIn = 50, numRuns = 200)
}
\arguments{
\item{dataframe}{data.frame usually taken from prep_corpus, a data.frame with fields for text, journal, date}

\item{n.topics}{numeric number of topics that we want from LDA}

\item{idcolname}{string the name of the id column from dataframe}

\item{textcolname}{string the name of the text column from dataframe}

\item{datecolname}{string the name of the date column from dataframe}

\item{stopListFile}{string location of the stop list file to be run with Mallet. Defaults to the
blank file provided by this package. Note that this file must have a single word per line,
and no special characters, and MUST be a .txt file}

\item{optFreq}{numeric how often should mallet optimize parameters?}

\item{burnIn}{numeric number of burn in iterations for LDA}

\item{numRuns}{numeric total number of runs Mallet should run before finalizing a model}
}
\value{
a Mallet object that from which we can get word frequencies, and can run LDA on
}
\description{
Create a Mallet object; from here we can gather word frequency statistics, and feed into createLDAvis
}
\examples{
date_vec <- date_vec
dataframe <- stripped_xmlData
marsden_journals <- make_model(dataframe, 10, journalVec = c("Blast", "Egoist", "Freewoman", "NewFreewoman"))
old_poetryMagazine <- make_model(dataframe, 10, journalVec = "Poetry Magazine", yearRangeRule = "< 1919")
poetry_vocab <- old_poetryMagazine$getVocabulary()
poetry_word.freqs <- mallet.word.freqs(marsden_journals)
}
\seealso{
\code{\link{prep_corpus}}, which creates the data.frame this function draws on
}

